Add files via upload
Explore the dataset and extract insights using Exploratory Data Analysis. Do the users spend more time on the new landing page than the old landing page? Is the conversion rate (the proportion of users who visit the landing page and get converted) for the new page greater than the conversion rate for the old page? Does the converted status depend on the preferred language?